montecarlo/dashboard_streamlit/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.22.0
kaleido>=0.2.1
streamlit>=1.36.0
scikit-learn>=1.3.0
python-dateutil>=2.8.2
